Создает лист комментов в popup
