+++ src/mock/film.js
@@ -1,7 +1,6 @@
 import {getRandomInteger} from '../utils';
 import {getRandomFloat} from '../utils';
-import dayjs from 'dayjs';
-
+import {createComment} from './comment';
 
 const films = [
   {
@@ -38,30 +37,13 @@ const countries = ['Finland', 'USA', 'Russia', 'Italy', 'France', 'German'];
 const genres = ['Comedy', 'Mystery', ' Film-Noir', 'Drama', 'Musical', 'Detective'];
 const description = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra.'.split('.');
 const ages = [0, 6, 12, 16, 18];
-const commentsList = ['Awesome!', 'Interesting setting and a good cast.', 'Almost two hours? Seriously?', 'Boooooring('];
-const emotions = ['smile', 'sleeping', 'puke', 'angry'];
 
-
-const createCommentDate = () => {
-  const monthGap = 11;
-  const dayGap = 30;
-  return dayjs().subtract(getRandomInteger(0, monthGap), 'month').subtract(getRandomInteger(0, dayGap), 'day');
-};
-
-const createComment = () => ({
-  id: getRandomInteger(0, 1000),
-  author: names[getRandomInteger(0, names.length - 1)],
-  comment: commentsList[getRandomInteger(0, commentsList.length - 1)],
-  date: createCommentDate(),
-  emotion: emotions[getRandomInteger(0, emotions.length - 1)],
-});
 
 export const createFilm = () => {
   const film = films[getRandomInteger(0, films.length - 1)];
   const watched = Boolean(getRandomInteger(0, 1));
   const watchedDate = watched ? new Date(getRandomInteger(2000, 2020), getRandomInteger(0, 11), getRandomInteger(0, 30)) : null;
-  const commentCount = getRandomInteger(0, 5);
-  const comments = new Array(commentCount).fill().map(() => createComment().id);
+  const comments = new Array(getRandomInteger(0, 5)).fill().map(createComment);
   return {
     comments: comments,
     userInfo: {
